Desafio-1/requirements.txt
# server/requirements.txt
flask
python-dotenv
requests
transformers
torch